step3  3️⃣ 로그인 리팩터링 (#49)
* test : add Test for FormAuthenticationConverter

* feat : implement convert method

* test : add convert test for BasicAuthenticationConverter

* feat : implement convert

* test : add test for NewSessionAuthenticationInterceptor

* feat : implement authentication

* test : write test for 'NewTokenAuthenticationInterceptor'

* feat : extract superclass and implement NewTokenAuthenticationInterceptor

* test : test if refactored codes pass acceptance tests

* refactor : remove old codes

* refactor : extract superclass

* refactor : disconnect circular dependency

* feat : add @AuthUser ArgumentResolver

* test : add test for 'MemberService'

* test : add 'findByEmail' test

* feat : implement findByEmail

* feart : change to use @AuthUser to get Logged In User

* test: add permission check test

* feat : implement permission check

+++ src/main/java/nextstep/subway/member/application/MemberService.java
@@ -1,11 +1,19 @@
 package nextstep.subway.member.application;
 
+import nextstep.subway.auth.domain.Authentication;
+import nextstep.subway.auth.domain.User;
+import nextstep.subway.auth.exception.AuthorizationException;
+import nextstep.subway.auth.infrastructure.SecurityContext;
+import nextstep.subway.auth.infrastructure.SecurityContextHolder;
+import nextstep.subway.member.domain.LoginMember;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
 import nextstep.subway.member.dto.MemberRequest;
 import nextstep.subway.member.dto.MemberResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Objects;
 
 @Service
 @Transactional
@@ -21,17 +29,39 @@ public class MemberService {
         return MemberResponse.of(member);
     }
 
+    @Transactional(readOnly = true)
     public MemberResponse findMember(Long id) {
         Member member = memberRepository.findById(id).orElseThrow(RuntimeException::new);
         return MemberResponse.of(member);
     }
 
     public void updateMember(Long id, MemberRequest param) {
+        checkPermission(id);
         Member member = memberRepository.findById(id).orElseThrow(RuntimeException::new);
         member.update(param.toMember());
     }
 
     public void deleteMember(Long id) {
+        checkPermission(id);
         memberRepository.deleteById(id);
+    }
+
+    @Transactional(readOnly = true)
+    public MemberResponse findMemberByEmail(String email) {
+        Member member = memberRepository.findByEmail(email).orElseThrow(RuntimeException::new);
+        return MemberResponse.of(member);
+    }
+
+    private void checkPermission(Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (Objects.isNull(authentication)) {
+            throw new AuthorizationException();
+        }
+
+        User user = (User) authentication.getPrincipal();
+
+        if (!Objects.equals(id, user.getId())) {
+            throw new AuthorizationException();
+        }
     }
 }
